feat: Interpret: Higher order functions
This took a while.

I had to rethink how the `Environment` worked.

Now it's a tree structure, with clones coming from a copied environment
of the current invocation of the function.

And I had to do something strange for recursive functions. Basically
inject the function value into the new function's environment. This is
because of the copy, I think.